requirements.txt
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
scikit-learn==1.7.2
spacy==3.8.11
torch==2.10.0
transformers==5.0.0
python-dotenv==1.1.0
PyYAML==6.0.2
